refactor : 반복 기능 관련 패키지 명 변경 및 위치 변경 #130
