+++ src/main/java/com/todoay/api/domain/todo/service/DailyTodoCRUDServiceImpl.java
@@ -16,8 +16,8 @@ import com.todoay.api.domain.todo.exception.NotYourTodoException;
 import com.todoay.api.domain.todo.exception.TodoNotFoundException;
 import com.todoay.api.domain.todo.repository.DailyTodoRepository;
 import com.todoay.api.domain.todo.utility.HashtagAttacher;
-import com.todoay.api.domain.todo.utility.RepeatType;
-import com.todoay.api.domain.todo.utility.Duration;
+import com.todoay.api.domain.todo.utility.repeat.RepeatType;
+import com.todoay.api.domain.todo.utility.repeat.Duration;
 import com.todoay.api.global.context.LoginAuthContext;
 import com.todoay.api.global.jwt.JwtProvider;
 import lombok.RequiredArgsConstructor;
