+++ src/main/java/com/todoay/api/domain/todo/utility/repeat/Duration.java
@@ -1,12 +1,10 @@
-package com.todoay.api.domain.todo.utility;
+package com.todoay.api.domain.todo.utility.repeat;
 
 
-import com.todoay.api.domain.todo.utility.inter.DateRepeator;
-import com.todoay.api.domain.todo.utility.inter.DateSelector;
-import com.todoay.api.domain.todo.utility.inter.select.CustomMonthDateSelector;
-import com.todoay.api.domain.todo.utility.inter.select.CustomNumberDateSelector;
-import com.todoay.api.domain.todo.utility.inter.select.OneMonthDateSelector;
-import com.todoay.api.domain.todo.utility.inter.select.CustomWeeksDateSelector;
+import com.todoay.api.domain.todo.utility.repeat.impl.selector.CustomMonthDateSelector;
+import com.todoay.api.domain.todo.utility.repeat.impl.selector.CustomNumberDateSelector;
+import com.todoay.api.domain.todo.utility.repeat.impl.selector.OneMonthDateSelector;
+import com.todoay.api.domain.todo.utility.repeat.impl.selector.CustomWeeksDateSelector;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
