+++ src/main/java/com/todoay/api/domain/todo/utility/repeat/impl/selector/CustomWeeksDateSelector.java
@@ -1,8 +1,8 @@
-package com.todoay.api.domain.todo.utility.inter.select;
+package com.todoay.api.domain.todo.utility.repeat.impl.selector;
 
 
-import com.todoay.api.domain.todo.utility.inter.DateRepeator;
-import com.todoay.api.domain.todo.utility.inter.DateSelector;
+import com.todoay.api.domain.todo.utility.repeat.DateRepeator;
+import com.todoay.api.domain.todo.utility.repeat.DateSelector;
 
 import java.time.LocalDate;
 import java.util.List;
